+++ src/miniJava/AbstractSyntaxTrees/Declaration.java
@@ -9,6 +9,9 @@ import miniJava.SyntacticAnalyzer.SourcePosition;
 
 public abstract class Declaration extends AST {
 	
+	public String name;
+	public TypeDenoter type;
+	
 	public Declaration(String name, TypeDenoter type) {
 		this(name, type, SourcePosition.ZERO);
 	}
@@ -17,8 +20,17 @@ public abstract class Declaration extends AST {
 		super(posn);
 		this.name = name;
 		this.type = type;
+		
+		// Error flags
+		this.duplicate_error = false;
+		this.selfref_error = false;
+		this.danglingdef_error = false;
 	}
 	
-	public String name;
-	public TypeDenoter type;
+	public abstract Declaration getMember(String name);
+	
+	// Error flags
+	public boolean duplicate_error;
+	public boolean selfref_error;
+	public boolean danglingdef_error;
 }