+++ src/miniJava/AbstractSyntaxTrees/FieldDecl.java
@@ -8,21 +8,25 @@ package miniJava.AbstractSyntaxTrees;
 import miniJava.SyntacticAnalyzer.SourcePosition;
 
 public class FieldDecl extends MemberDecl {
-	
+
 	public FieldDecl(boolean isPrivate, boolean isStatic, TypeDenoter t, String name) {
 		this(isPrivate, isStatic, t, name, SourcePosition.ZERO);
 	}
-	
-	public FieldDecl(boolean isPrivate, boolean isStatic, TypeDenoter t, String name, SourcePosition posn){
-    super(isPrivate, isStatic, t, name, posn);
-	}
-	
-	public FieldDecl(MemberDecl md, SourcePosition posn) {
-		super(md,posn);
-	}
-	
-	public <A, R> R visit(Visitor<A, R> v, A o) {
-        return v.visitFieldDecl(this, o);
-    }
-}
 
+	public FieldDecl(boolean isPrivate, boolean isStatic, TypeDenoter t, String name, SourcePosition posn) {
+		super(isPrivate, isStatic, t, name, posn);
+	}
+
+	public FieldDecl(MemberDecl md, SourcePosition posn) {
+		super(md, posn);
+	}
+
+	public <A, R> R visit(Visitor<A, R> v, A o) {
+		return v.visitFieldDecl(this, o);
+	}
+
+	@Override
+	public Declaration getMember(String name) {
+		return this.type.getMember(name);
+	}
+}