+++ src/miniJava/AbstractSyntaxTrees/Identifier.java
@@ -8,13 +8,22 @@ package miniJava.AbstractSyntaxTrees;
 import miniJava.SyntacticAnalyzer.Token;
 
 public class Identifier extends Terminal {
+	private Declaration decl;
 
-  public Identifier (Token t) {
-    super (t);
-  }
+	public Identifier(Token t) {
+		super(t);
+	}
 
-  public <A,R> R visit(Visitor<A,R> v, A o) {
-      return v.visitIdentifier(this, o);
-  }
+	public <A, R> R visit(Visitor<A, R> v, A o) {
+		return v.visitIdentifier(this, o);
+	}
+	
+	public void setDecl(Declaration decl) {
+		this.decl = decl;
+	}
+	
+	public Declaration getDecl() {
+		return this.decl;
+	}
 
 }