+++ src/miniJava/AbstractSyntaxTrees/changes.md
@@ -1,1 +1,14 @@
+### Declaration
+
+ * Added duplicate_error flag
+ * Added selfref_error flag
+
+### Reference
+
+ * Added illegal_nonstatic_error flag
+
+### Class Decl
+
+ * Add 'Scope' field
+
 ### NullLiteral